move decorators to draft/shared
